Update README.md
Embedded Systems Developer/Robotics Enthusiast
Welcome to my GitHub repository! This repository is an idea that started in my final year of undergraduate studies, and it also serves as a showcase of my projects and skills as an aspiring embedded developer and/or robotic engineer. Whether you're a fellow developer, a potential employer, or just curious about embedded systems, I hope you find something interesting here and feel free to contribute to improving this idea.
About Me
I am a passionate and dedicated master’s studies student, pursuing a career in Intelligent embedded systems development. My background is in Computer Science, with a strong foundation in programming, electronics, and embedded systems design, and I am a member of https://github.com/Intelligent-Systems-DVA439 , an active community of Intelligent Systems enthusiasts, where we share knowledge as well as collaborate on various projects. I am constantly seeking opportunities to enhance my skills and contribute to innovative projects in embedded systems and robotics fields.